MISUV-3590:   * refactor validMFACreditDescription method into ~Utils as its shared between Payment and DocumentDetails models;   * add documentDescription and validation method to Payment model;
