+++ app/models/financialDetails/PaymentsResponse.scala
@@ -27,6 +27,7 @@ case class PaymentsError(status: Int, error: String) extends PaymentsResponse
 case class Payment(reference: Option[String],
                    amount: Option[BigDecimal],
                    method: Option[String],
+                   documentDescription: Option[String],
                    lot: Option[String],
                    lotItem: Option[String],
                    date: Option[String],
@@ -39,6 +40,7 @@ case class Payment(reference: Option[String],
     case Some(credit) => Some(credit)
   }
 
+  def validMFACreditDescription() : Boolean = MfaCreditUtils.validMFACreditDescription(this.documentDescription)
 }
